requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.5
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
flask_postgres==0.2.2
greenlet==2.0.2
idna==3.4
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.18
SQLAlchemy-Utils==0.41.1
typing_extensions==4.7.1
validators==0.20.0
Werkzeug==2.3.6
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
